Second wave of changes addressing second PR review
